provides 'rebuild loader' ability
The rebuild checks for user's intention by way
of standard @Generated annotaton.  If the @Generated annotation is
absent, the rebuild will preserve the marked java element: method or field.

The user can make changes to the generator's output....and can
expressly preserve that change by removing the @Generated from the
modified element.