generate inline statements in methods, preserve user defined methods
The goals is to avoid losing customizations among the forest
of generated methods.

Prior, generator always created 4 field methods for every field:
with${field}, from${field}, modify${field} and initialize${field}.
Now, only one:  with${field} is still generated.

Prior, the loader methods from, modify and intialize would call
the above field methods, respectively, for each field.  Now,
instead of calling these methods, the statements are inlined to the
loader methods.

Now, the generator detects the presense of user defined field methods.
If present, the generator writes statements to call them from
respective loader methods ('from' loader method calles 'from${field}'
method).  The generator writes an 'inline' statement otherwise.

The reason for all of the above, is so there is less code in the
Loader from the generator.  And still, the developer can customize
behavior for each field.

Prior, the developer would customize a generated field method and remove the
@Generated.  Now, the developer would originate a field method,
and the generator will detect it and invoke generate a call
by the respective loader method.